fix: add compress command
